Take in String instead of &str to sanitize_rfc822_like_date() and related functions
Didn't do so for parse_from_rfc2822_with_fallback() ebcause (a) we
don't always move and (b) it would just be confusing

But back to the point. The previous code would allocate an insane amount
of Strings that were immediately discarded. This one will allocate *0*
in best-case and linear n for n=amount of corrections made